Improve matrix
